+++ packages/dataviews/src/components/dataviews/index.tsx
@@ -13,7 +13,7 @@ import { useResizeObserver } from '@wordpress/compose';
 /**
  * Internal dependencies
  */
-import DataViewsContext, { DEFAULT_VIEW_EMPTY } from '../dataviews-context';
+import DataViewsContext from '../dataviews-context';
 import {
 	default as DataViewsFilters,
 	useFilters,
@@ -23,20 +23,13 @@ import DataViewsLayout from '../dataviews-layout';
 import DataViewsFooter from '../dataviews-footer';
 import DataViewsSearch from '../dataviews-search';
 import { BulkActionsFooter } from '../dataviews-bulk-actions';
-import { DataViewsEmpty } from '../dataviews-empty';
 import { DataViewsPagination } from '../dataviews-pagination';
 import DataViewsViewConfig, {
 	DataviewsViewConfigDropdown,
 	ViewTypeMenu,
 } from '../dataviews-view-config';
 import { normalizeFields } from '../../normalize-fields';
-import type {
-	Action,
-	Field,
-	View,
-	SupportedLayouts,
-	EmptyViewProps,
-} from '../../types';
+import type { Action, Field, View, SupportedLayouts } from '../../types';
 import type { SelectionOrUpdater } from '../../private-types';
 type ItemWithId = { id: string };
 
@@ -67,7 +60,7 @@ type DataViewsProps< Item > = {
 	getItemLevel?: ( item: Item ) => number;
 	children?: ReactNode;
 	perPageSizes?: number[];
-	empty?: EmptyViewProps;
+	empty?: ReactNode;
 } & ( Item extends ItemWithId
 	? { getItemId?: ( item: Item ) => string }
 	: { getItemId: ( item: Item ) => string } );
@@ -208,7 +201,7 @@ function DataViews< Item >( {
 				isShowingFilter,
 				setIsShowingFilter,
 				perPageSizes,
-				empty: empty ?? DEFAULT_VIEW_EMPTY,
+				empty,
 			} }
 		>
 			<div className="dataviews-wrapper" ref={ containerRef }>
@@ -234,7 +227,6 @@ const DataViewsSubComponents = DataViews as typeof DataViews & {
 	Pagination: typeof DataViewsPagination;
 	Search: typeof DataViewsSearch;
 	ViewConfig: typeof DataviewsViewConfigDropdown;
-	Empty: typeof DataViewsEmpty;
 };
 
 DataViewsSubComponents.BulkActionToolbar = BulkActionsFooter;
@@ -245,6 +237,5 @@ DataViewsSubComponents.LayoutSwitcher = ViewTypeMenu;
 DataViewsSubComponents.Pagination = DataViewsPagination;
 DataViewsSubComponents.Search = DataViewsSearch;
 DataViewsSubComponents.ViewConfig = DataviewsViewConfigDropdown;
-DataViewsSubComponents.Empty = DataViewsEmpty;
 
 export default DataViewsSubComponents;